requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
async-timeout==3.0.1
attrs==23.1.0
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
frozenlist==1.3.3
h11==0.14.0
idna==3.4
multidict==6.0.4
py-cord==2.4.1
python-dotenv==1.0.0
typing-extensions==4.6.0
uvicorn==0.22.0
yarl==1.9.2
